msm_rng: Change long to u32 type for buffers reading from RNG HW
Changing the type from long to u32 of the buffer that is used to
read data from the RNG hardware to reflect the size of the HW
register

Acked-by: Baranidharan Muthukumaran <bmuthuku@qti.qualcomm.com>
(cherry picked from commit 7bcafc734de9cfd661bdd6da9e997123c821e130)

+++ drivers/char/hw_random/msm_rng.c
@@ -56,8 +56,8 @@ static int msm_rng_read(struct hwrng *rng, void *data, size_t max, bool wait)
 	void __iomem *base;
 	size_t maxsize;
 	size_t currsize = 0;
-	unsigned long val;
-	unsigned long *retdata = data;
+	u32 val;
+	u32 *retdata = data;
 	int ret;
 
 	msm_rng_dev = (struct msm_rng_device *)rng->priv;
